+++ app/components/Footer/route.ts
